Created search as ajax request

+++ sentiment/views.py
@@ -2,34 +2,35 @@ from django.shortcuts import render
 from django.views.decorators.csrf import csrf_exempt
 from .sentiment_analysis.analysis import getDataToView
 from .models import news
+from django.http import HttpResponseRedirect, JsonResponse, HttpResponse
 
 # Create your views here.
+@csrf_exempt
 def index(request):
+    if request.is_ajax():
+        print(request.POST['search'])
+        news.objects.all().delete()
+        try:
+            NewsData = getDataToView(request.POST['search'])
+            for data in NewsData:
+                newsObject = news(
+                    newstitle = data['title'],
+                    newsdescription = data['description'],
+                    date = data['date'],
+                    source = data['source'],
+                    negative_value = data['negative'],
+                    neutral_value = data['neutral'],
+                    positive = data['positive']
+                )
+                newsObject.save()
+
+            return JsonResponse({'insert': 'success'})
+        
+        except:
+            pass
+
     return render(request, 'index.html')
 
 @csrf_exempt
 def chart(request):
-    news.objects.all().delete()
-    try:
-        NewsData = getDataToView(request.POST['title'])
-        for data in NewsData:
-            newsObject = news(
-                newstitle = data['title'],
-                newsdescription = data['description'],
-                date = data['date'],
-                source = data['source'],
-                negative_value = data['negative'],
-                neutral_value = data['neutral'],
-                positive = data['positive']
-            )
-
-            newsObject.save()
-
-        return render(request, 'chart.html')
-
-    except:
-        return render(request, 'error.html')
-
-
-def visualize(request):
-    news
+    return render(request, 'chart.html')
